D3.b Complete. Cleaned up comments, Caches and playerLocation use the Flyweight pattern correctly.
After I felt like I was finished, I had Brace double-check my implementation of the Flyweight pattern and it suggest to modify Cache's so that they have a Cell as part of them, rather than extending the Cell object to prevent duplication. After cleaning up the code, I made the modification and now it is ready to submit.